+++ src/core/request/engines/provider/index.ts
@@ -67,9 +67,6 @@ export default function createProviderEngine(
 			p = <AvailableOptions>Object.select(params, availableParams),
 			provider = getProviderInstance(src, p.meta);
 
-		p.query = restoreEnumerable(p.query);
-		p.body = restoreEnumerable(p.body);
-
 		const defaultRequestMethods = providerMethodProperties.reduceRight((carry, key) => {
 			const
 				method = provider[key];
@@ -214,28 +211,4 @@ function createMixedProvider(base: Provider, modifier: Provider | Dictionary<str
 	});
 
 	return mixedProvider;
-}
-
-/**
- * Restores enumerable option (sets true) for properties of an object
- * @param obj - an object for the restoration
- */
-function restoreEnumerable<T extends Dictionary | unknown>(obj: T): T {
-	if (!Object.isPlainObject(obj)) {
-		return obj;
-	}
-
-	const
-		keys = Object.keys(obj),
-		names = Object.getOwnPropertyNames(obj);
-
-	if (keys.length === names.length) {
-		return obj;
-	}
-
-	return names.reduce((carry, key) => {
-		carry[key] = obj[key];
-
-		return carry;
-	}, Object.create(obj));
 }